pyproject.toml
[tool.poetry]
name = "mm-plugins"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
"discord.py" = "^2.3.2"
beaupy = "^3.7.2"


[tool.poetry.group.dev.dependencies]
pylint = "^3.0.3"
autopep8 = "^2.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
